Extracts spy's to test package, and interfaces to main package
