+++ src/main/java/HttpRequest.java
@@ -1,0 +1,3 @@
+class HttpRequest {
+
+}
