+++ src/main/java/HttpRequestParser.java
@@ -1,0 +1,7 @@
+import java.io.InputStream;
+
+interface HttpRequestParser {
+
+    HttpRequest parse(InputStream inputStream);
+
+}
