+++ src/main/java/HttpServerSocket.java
@@ -1,0 +1,3 @@
+public interface HttpServerSocket {
+    HttpSocket accept();
+}
