+++ src/main/java/HttpSocket.java
@@ -1,0 +1,5 @@
+import java.io.InputStream;
+
+public interface HttpSocket {
+    InputStream getRawHttpRequest();
+}
